requirements.txt
faiss-cpu==1.11.0
langchain==0.3.25
langchain-community==0.3.24
langchain-google-genai==2.1.5
langchain-huggingface==0.2.0
PyMuPDF==1.26.0
python-dotenv==1.1.0